Refactor implementations into games.js
